created _BaseIT and extended it for all integration tests
