+++ src/test/java/com/pbuczek/pf/it/EncounterIT.java
@@ -1,92 +1,48 @@
 package com.pbuczek.pf.it;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.ObjectWriter;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.pbuczek.pf.TestUserDetails;
 import com.pbuczek.pf.encounter.Encounter;
 import com.pbuczek.pf.encounter.EncounterDto;
 import com.pbuczek.pf.encounter.EncounterRepository;
-import com.pbuczek.pf.user.User;
-import com.pbuczek.pf.user.UserDto;
-import com.pbuczek.pf.user.UserRepository;
-import com.pbuczek.pf.user.UserType;
 import lombok.SneakyThrows;
 import org.apache.commons.lang3.RandomStringUtils;
-import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.actuate.observability.AutoConfigureObservability;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockHttpServletResponse;
-import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.ResultActions;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.Base64;
 import java.util.List;
 import java.util.Optional;
 
 import static com.pbuczek.pf.encounter.Encounter.MAX_DESCRIPTION_LENGTH;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.junit.jupiter.api.Assertions.fail;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @Tag("IntegrationTest")
-@AutoConfigureObservability
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@AutoConfigureMockMvc
-class EncounterIT implements TestUserDetails {
+class EncounterIT extends _BaseIT {
 
-    private final List<Integer> createdUserIds = new ArrayList<>();
     private final List<Integer> createdEncounterIds = new ArrayList<>();
-    private static final ObjectMapper mapper = new ObjectMapper();
-    private static ObjectWriter ow;
     private static final String encName = "testEncounterName";
 
     @Autowired
-    private MockMvc mockMvc;
-    @Autowired
-    private UserRepository userRepo;
-    @Autowired
     private EncounterRepository encounterRepo;
-
-    @BeforeAll
-    static void initialize() {
-        mapper.configure(SerializationFeature.WRAP_ROOT_VALUE, false);
-        mapper.registerModule(new JavaTimeModule());
-        ow = mapper.writer().withDefaultPrettyPrinter();
-    }
-
-    @BeforeEach
-    void setUp() {
-        Integer potentialId = userRepo.getIdByUsername(TEST_USERNAME_ADMIN_1);
-        if (potentialId != null) {
-            userRepo.deleteUser(potentialId);
-        }
-        User admin = new User(TEST_USERNAME_ADMIN_1, TEST_EMAIL_ADMIN_1, TEST_PASSWORD);
-        admin.setType(UserType.ADMIN);
-        admin.setEnabled(true);
-        userRepo.save(admin);
-        createdUserIds.add(admin.getId());
-    }
 
     @AfterEach
     void tearDown() {
         createdEncounterIds.forEach(id -> encounterRepo.deleteEncounter(id));
-        createdUserIds.forEach(id -> userRepo.deleteUser(id));
     }
 
     @Test
     @SneakyThrows
     void encounterIsCreatedCorrectly() {
-        int userId = createUserAndGetId(TEST_USERNAME_STANDARD_1, TEST_EMAIL_STANDARD_1);
+        int userId = createUser(TEST_USERNAME_STANDARD_1, TEST_EMAIL_STANDARD_1);
         MvcResult result = getResultActionsForCreatingEncounter(userId, "test")
                 .andExpect(status().isOk()).andReturn();
 
@@ -108,7 +64,7 @@ class EncounterIT implements TestUserDetails {
     @Test
     @SneakyThrows
     void cannotCreateEncounterWithDescriptionTooLong() {
-        int userId = createUserAndGetId(TEST_USERNAME_STANDARD_1, TEST_EMAIL_STANDARD_1);
+        int userId = createUser(TEST_USERNAME_STANDARD_1, TEST_EMAIL_STANDARD_1);
         MockHttpServletResponse response = getResultActionsForCreatingEncounter(userId,
                 RandomStringUtils.random(3001, true, true))
                 .andExpect(status().isBadRequest()).andReturn().getResponse();
@@ -120,7 +76,7 @@ class EncounterIT implements TestUserDetails {
     @Test
     @SneakyThrows
     void differentStandardUserCannotReadEncounterThatIsNotPublished() {
-        int userId = createUserAndGetId(TEST_USERNAME_STANDARD_1, TEST_EMAIL_STANDARD_1);
+        int userId = createUser(TEST_USERNAME_STANDARD_1, TEST_EMAIL_STANDARD_1);
         MvcResult postResult = getResultActionsForCreatingEncounter(userId, "test")
                 .andExpect(status().isOk()).andReturn();
 
@@ -128,7 +84,7 @@ class EncounterIT implements TestUserDetails {
         basicEncounterChecks(createdEncounter);
         Integer createdEncounterId = createdEncounter.getId();
 
-        int secondUserId = createUserAndGetId(TEST_USERNAME_STANDARD_2, TEST_EMAIL_STANDARD_2);
+        int secondUserId = createUser(TEST_USERNAME_STANDARD_2, TEST_EMAIL_STANDARD_2);
         enableUserAccount(secondUserId);
 
         assertThat(this.mockMvc.perform(
@@ -141,7 +97,7 @@ class EncounterIT implements TestUserDetails {
     @Test
     @SneakyThrows
     void differentStandardUserCanReadEncounterThatIsPublished() {
-        int userId = createUserAndGetId(TEST_USERNAME_STANDARD_1, TEST_EMAIL_STANDARD_1);
+        int userId = createUser(TEST_USERNAME_STANDARD_1, TEST_EMAIL_STANDARD_1);
         MvcResult postResult = getResultActionsForCreatingEncounter(userId, "test")
                 .andExpect(status().isOk()).andReturn();
 
@@ -149,7 +105,7 @@ class EncounterIT implements TestUserDetails {
         basicEncounterChecks(createdEncounter);
         Integer createdEncounterId = createdEncounter.getId();
 
-        int secondUserId = createUserAndGetId(TEST_USERNAME_STANDARD_2, TEST_EMAIL_STANDARD_2);
+        int secondUserId = createUser(TEST_USERNAME_STANDARD_2, TEST_EMAIL_STANDARD_2);
         enableUserAccount(secondUserId);
         publishUnpublishEncounter(createdEncounterId);
 
@@ -162,22 +118,6 @@ class EncounterIT implements TestUserDetails {
         assertThat(foundEncounter).usingRecursiveComparison()
                 .ignoringFields("published")
                 .isEqualTo(createdEncounter);
-    }
-
-    private int createUserAndGetId(String username, String email) throws Exception {
-        MockHttpServletResponse response = this.mockMvc.perform(
-                        post("/user")
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(ow.writeValueAsString(
-                                        new UserDto(username, email, TestUserDetails.TEST_PASSWORD))))
-                .andReturn().getResponse();
-        User user = mapper.readValue(response.getContentAsString(), User.class);
-        if (user.getId() != null) {
-            createdUserIds.add(user.getId());
-        } else {
-            fail("User could not be created correctly");
-        }
-        return user.getId();
     }
 
     @SneakyThrows
@@ -193,17 +133,6 @@ class EncounterIT implements TestUserDetails {
         assert createdEncounter != null;
         assertThat(createdEncounter.getId()).isNotNull();
         createdEncounterIds.add(createdEncounter.getId());
-    }
-
-    private String getBasicAuthenticationHeader(String username) {
-        String valueToEncode = username + ":" + TestUserDetails.TEST_PASSWORD;
-        return "Basic " + Base64.getEncoder().encodeToString(valueToEncode.getBytes());
-    }
-
-    private void enableUserAccount(int userId) throws Exception {
-        this.mockMvc.perform(
-                patch("/user/enable/" + userId)
-                        .header("Authorization", getBasicAuthenticationHeader(TEST_USERNAME_ADMIN_1)));
     }
 
     private void publishUnpublishEncounter(int encounterId) throws Exception {
