+++ src/test/java/com/pbuczek/pf/it/UserIT.java
@@ -1,73 +1,24 @@
 package com.pbuczek.pf.it;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.ObjectWriter;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.pbuczek.pf.TestUserDetails;
-import com.pbuczek.pf.user.*;
+import com.pbuczek.pf.user.PaymentPlan;
+import com.pbuczek.pf.user.User;
+import com.pbuczek.pf.user.UserType;
 import lombok.SneakyThrows;
 import org.apache.commons.lang3.RandomStringUtils;
-import org.junit.jupiter.api.*;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.actuate.observability.AutoConfigureObservability;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockHttpServletResponse;
-import org.springframework.test.web.servlet.MockMvc;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @Tag("IntegrationTest")
-@AutoConfigureObservability
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@AutoConfigureMockMvc
-class UserIT implements TestUserDetails {
-
-    private final List<Integer> createdUserIds = new ArrayList<>();
-    private static final ObjectMapper mapper = new ObjectMapper();
-    private static ObjectWriter ow;
-
-    @Autowired
-    private MockMvc mockMvc;
-    @Autowired
-    UserRepository userRepo;
-
-    @BeforeAll
-    static void initialize() {
-        mapper.configure(SerializationFeature.WRAP_ROOT_VALUE, false);
-        mapper.registerModule(new JavaTimeModule());
-        ow = mapper.writer().withDefaultPrettyPrinter();
-    }
-
-    @BeforeEach
-    void setUp() {
-        Integer potentialId = userRepo.getIdByUsername(TEST_USERNAME_ADMIN_1);
-        if (potentialId != null) {
-            userRepo.deleteUser(potentialId);
-        }
-        User admin = new User(TEST_USERNAME_ADMIN_1, TEST_EMAIL_ADMIN_1, TEST_PASSWORD);
-        admin.setType(UserType.ADMIN);
-        admin.setEnabled(true);
-        userRepo.save(admin);
-        createdUserIds.add(admin.getId());
-    }
-
-    @AfterEach
-    void tearDown() {
-        createdUserIds.forEach(id -> userRepo.deleteUser(id));
-    }
+class UserIT extends _BaseIT {
 
     @Test
     @SneakyThrows
@@ -131,26 +82,6 @@ class UserIT implements TestUserDetails {
 
         createUser(TEST_USERNAME_STANDARD_1, TEST_EMAIL_ADMIN_1, HttpStatus.CONFLICT,
                 String.format("email '%s' is already being used by another user.", TEST_EMAIL_ADMIN_1));
-    }
-
-    @SneakyThrows
-    private MockHttpServletResponse createUser(
-            String username, String email, HttpStatus expectedStatus) {
-
-        return this.mockMvc.perform(
-                        post("/user")
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(ow.writeValueAsString(
-                                        new UserDto(username, email, TestUserDetails.TEST_PASSWORD))))
-                .andExpect(status().is(expectedStatus.value())).andReturn().getResponse();
-    }
-
-    @SneakyThrows
-    private void createUser(
-            String username, String email, HttpStatus expectedStatus, String expectedErrorMessage) {
-
-        MockHttpServletResponse response = createUser(username, email, expectedStatus);
-        assertThat(response.getErrorMessage()).isEqualTo(expectedErrorMessage);
     }
 
     @SneakyThrows
